Attributes approaching completion

+++ src/main/java/org/devine/aberrant/advantage/AdvantageRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdvantageRepository extends JpaRepository<Advantage, Integer> {
+public interface AdvantageRepository extends JpaRepository<Advantage, Long> {
 }
