+++ src/main/java/org/devine/aberrant/attribute/AttributeRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AttributeRepository extends JpaRepository<Attribute, Integer> {
+public interface AttributeRepository extends JpaRepository<Attribute, Long> {
 }
