+++ src/main/java/org/devine/aberrant/background/BackgroundRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BackgroundRepository extends JpaRepository<Background, Integer> {
+public interface BackgroundRepository extends JpaRepository<Background, Long> {
 }
