+++ src/main/java/org/devine/aberrant/megaAttribute/MegaAttributeRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MegaAttributeRepository extends JpaRepository<MegaAttribute, Integer> {
+public interface MegaAttributeRepository extends JpaRepository<MegaAttribute, Long> {
 }
