+++ src/main/java/org/devine/aberrant/power/PowerRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PowerRepository extends JpaRepository<Power, Integer> {
+public interface PowerRepository extends JpaRepository<Power, Long> {
 }
